+++ kafka-connector-project/kafka-connector/src/test/java/com/lightstreamer/kafka/adapters/consumers/processor/RecordProcessorTest.java
@@ -24,6 +24,7 @@ import static com.lightstreamer.kafka.common.mapping.selectors.DataExtractor.ext
 import com.lightstreamer.kafka.adapters.consumers.processor.RecordConsumer.RecordProcessor;
 import com.lightstreamer.kafka.common.mapping.Items;
 import com.lightstreamer.kafka.common.mapping.Items.SubscribedItem;
+import com.lightstreamer.kafka.common.mapping.Items.SubscribedItems;
 import com.lightstreamer.kafka.common.mapping.RecordMapper;
 import com.lightstreamer.kafka.common.mapping.RecordMapper.Builder;
 import com.lightstreamer.kafka.common.mapping.selectors.ExtractionException;
@@ -37,6 +38,7 @@ import org.junit.jupiter.api.Test;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 
 public class RecordProcessorTest {
@@ -48,8 +50,11 @@ public class RecordProcessorTest {
     }
 
     private RecordMapper<String, String> mapper;
+
     private AtomicInteger counter;
+    private AtomicBoolean snapshotEvent;
     private MockItemEventListener listener;
+
     private ConsumerRecord<String, String> record;
     private Set<SubscribedItem> subscribedItems;
     private RecordProcessor<String, String> processor;
@@ -68,11 +73,19 @@ public class RecordProcessorTest {
         // Counts the listener invocations to deliver the real-time updates
         this.counter = new AtomicInteger();
 
-        // The mocked ItemEventListener instance, which updates the counter upon invocation.
-        this.listener = new MockItemEventListener(update -> counter.incrementAndGet());
+        // Indicates whether the delivered event is a snapshot
+        this.snapshotEvent = new AtomicBoolean(false);
+
+        // The mocked ItemEventListener instance, which updates the counter upon invocation
+        this.listener =
+                new MockItemEventListener(
+                        (update, isSnapshot) -> {
+                            counter.incrementAndGet();
+                            snapshotEvent.set(isSnapshot);
+                        });
 
         // A record routable to "item1" and "item2"
-        this.record = Records.Record(TEST_TOPIC, 0, "a-1");
+        this.record = Records.ConsumerRecord(TEST_TOPIC, 0, "a-1");
 
         // The collection of subscribable items
         this.subscribedItems = new HashSet<>();
@@ -80,7 +93,7 @@ public class RecordProcessorTest {
         // The RecordProcessor instance
         this.processor =
                 new RecordConsumerSupport.DefaultRecordProcessor<>(
-                        mapper, subscribedItems, listener);
+                        mapper, SubscribedItems.of(subscribedItems), listener);
     }
 
     @Test
@@ -88,14 +101,21 @@ public class RecordProcessorTest {
         // Subscribe to "item1" and process the record
         subscribedItems.add(Items.subscribedFrom("item1", new Object()));
         processor.process(record);
+
         // Verify that the real-time update has been routed
         assertThat(counter.get()).isEqualTo(1);
+        // Verify that the update has NOT been routed as a snapshot
+        assertThat(snapshotEvent.get()).isFalse();
+
         // Reset the counter
         counter.set(0);
+        // Reset the snapshot flag
+        snapshotEvent.set(false);
 
-        // Add subscription "item2"
+        // Add subscription "item2" and process the record
         subscribedItems.add(Items.subscribedFrom("item2", new Object()));
         processor.process(record);
+
         // Verify that the update has been routed two times, one for "item1" and one for "item2"
         assertThat(counter.get()).isEqualTo(2);
     }
@@ -107,5 +127,10 @@ public class RecordProcessorTest {
         processor.process(record);
         // Verify that the update has NOT been routed
         assertThat(counter.get()).isEqualTo(0);
+    }
+
+    @Test
+    public void shouldNotBeCommandEnforceEnabled() {
+        assertThat(processor.isCommandEnforceEnabled()).isFalse();
     }
 }